move Event implementation for Hashmap to event_json
