+++ sds/src/lib.rs
@@ -7,6 +7,8 @@ mod encoding;
 mod event;
 mod match_action;
 
+#[cfg(any(test, feature = "testing", feature = "bench"))]
+mod event_json;
 mod match_validation;
 mod normalization;
 mod observability;
@@ -17,11 +19,10 @@ mod rule_match;
 mod scanner;
 mod scoped_ruleset;
 mod secondary_validation;
-mod stats;
-mod validation;
-
 #[cfg(any(test, feature = "testing", feature = "bench"))]
 mod simple_event;
+mod stats;
+mod validation;
 
 // This is the public API of the SDS core library
 pub use encoding::{EncodeIndices, Encoding, Utf8Encoding};
